Update the init imports to be correct, rename the group_event file to group_event_handlers to further disambiguate between it and grouptype.py
